Fix easyQC code CHCP FAMILY ABCD

+++ QC/CHCP/CHCP_RunEasyQC2.r
@@ -21,67 +21,76 @@ library(R.utils)
 # Prepare files 
 
 
-  for (pheni in "maxRCF"){
-    for (sexi in c("MW","WOMEN", "MEN" )){
+  for (pheni in c("maxRCF", "meanRCF", "minRCF",  "maxRCN", "meanRCN", "minRCN")){ #
+    for (sexi in c("MW","MEN", "WOMEN")){ # "MW", 
       for (adji in c("noBMI", "adjBMI") ){
-        for(popi in c("EUR")){
-          for (agegri in "ADL"){
+        for(popi in c("EUR", "nonEUR")){ # , "nonEUR"
+          for (agegri in c("ADO", "ADL", "OAD")){# , "AAD"
 
             filei <- paste0("/CP_cIMT.", pheni, ".", agegri,".",sexi,".",adji,".",popi, ".KL.230922")
-             
+            if1<- paste0(ir_p,"/Raw/Modified/", filei, ".txt")
+            if2<- paste0(ir_p,"/Raw/Modified/", filei, "_2.txt")
+            if3<- paste0(ir_p,"/Raw/Modified/", filei, "_3.txt")
 
-            if1<- paste0(ir_p,"/Raw/Modified/",pheni, "/",agegri,"/", filei, ".txt")
-            if2<- paste0(ir_p,"/Raw/Modified/",pheni, "/",agegri,"/", filei, "_2.txt")
-            
-            if(!file.exists(paste0(if2, ".gz"))){
+            if(!file.exists(paste0(if3, ".gz"))){
               
-              if(!file.exists(if1) ){
-                unzip( paste0(ir_p,"/Raw/", filei, ".zip"), exdir=paste0(ir_p,"/Raw/Modified/",pheni, "/",agegri,"/") )
+              cat (" \n reading file", pheni, sexi, adji, popi, agegri, "\n") 
+              if(file.exists(paste0(if2, ".gz"))){
+
+               chcp <- as.data.frame(fread(paste0(if2, ".gz")))
+              } else {
+                if(!file.exists(if1) ){ 
+                  unzip( paste0(ir_p,"/Raw/", filei, ".zip"), exdir=paste0(ir_p,"/Raw/Modified/") )
+                  chcp <- as.data.frame(fread(paste0(if1, ".gz")))
+                }
+                if (file.exists(if1)){
+                  chcp <- as.data.frame(fread(paste0(if1)))
+                }
               }
               
-              cat (" \n reading file \n") 
-              chcp <- fread(paste0(if1), head=T)
-  
-              cat (" \n saving file \n") 
-              write.table(chcp,  if2, row.names=F, col.names=T, quote=F, sep="\t" )
-              
-              cat (" \n zipping file \n") 
-
-              gzip(if2, ext="gz", FUN=gzfile)
-              file.remove(if1)
+              chcp[which(chcp$INFO_TYPE==0),"INFO"] <- 1
+              cat (" \n            saving file \n") 
+              write.table(chcp, if3, row.names=F, col.names=T, quote=F, sep="\t" )
+              cat (" \n            zipping file \n") 
+              gzip(if3, ext="gz", FUN=gzfile)
+              if(file.exists(if1))file.remove(if1)
 
             }
-
           } 
         }
       }
     }
   }
 
-
-
-
-
-
 # Run EasyQC 
-  # ADO maxRCF EUR 
-    setwd(paste0(o_p, "/ADO/maxRCF/EUR"))
-    EasyQC2(paste0(qc2_p, "/CHCP_ADO_maxRCF_EUR_QC_quant.ecf"))
-     	
-  # OAD  EUR
-    setwd(paste0(o_p, "/OAD/maxRCF/nonEUR"))
-    EasyQC2(paste0(qc2_p, "//CHCP/EGG_cIMT_CHCP_maxRCF_OAD_nonEUR_QC_quant.ecf"))
+  # maxRCF 
+    # EUR
+      # ADO 
+        setwd(paste0(o_p, "/ADO/maxRCF/EUR"))
+        EasyQC2(paste0(qc2_p, "/CHCP_ADO_maxRCF_EUR_QC_quant.ecf"))
+    
+      # OAD
+        setwd(paste0(o_p, "/OAD/maxRCF/EUR"))
+        EasyQC2(paste0(qc2_p, "/CHCP_OAD_maxRCF_EUR_QC_quant.ecf"))
+      	
+      # ADL
+        setwd(paste0(o_p, "/ADL/maxRCF/EUR/"))
+        EasyQC2(paste0(qc2_p, "/CHCP_ADL_maxRCF_EUR_QC_quant.ecf"))
+    
 
-  # OAD  nonEUR
-    setwd(paste0(o_p, "/OAD/maxRCF/nonEUR"))
-    EasyQC2(paste0(qc2_p, "//CHCP/EGG_cIMT_CHCP_maxRCF_OAD_nonEUR_QC_quant.ecf"))
 
-  	
-  # ADL  EUR
-    setwd(paste0(o_p, "/ADL/EUR/maxRCF/"))
-    EasyQC2(paste0(qc2_p, "//CHCP/EGG_cIMT_CHCP_maxRCF_ADL_EUR_QC_quant.ecf"))
 
-  # ADL  nonEUR
-    setwd(paste0(o_p, "/ADL/maxRCF/nonEUR"))
-    EasyQC2(paste0(qc2_p, "//CHCP/EGG_cIMT_CHCP_maxRCF_ADL_nonEUR_QC_quant.ecf"))
+    
+    # nonEUR
 
+      # OAD 
+        setwd(paste0(o_p, "/OAD/maxRCF/nonEUR"))
+        EasyQC2(paste0(qc2_p, "/CHCP_OAD_maxRCF_nonEUR_QC_quant.ecf"))
+
+      # ADL  nonEUR
+        setwd(paste0(o_p, "/ADL/maxRCF/nonEUR"))
+        EasyQC2(paste0(qc2_p, "//CHCP_ADL_maxRCF_nonEUR_QC_quant.ecf"))  
+
+        # ADO
+        setwd(paste0(o_p, "/ADO/maxRCF/nonEUR"))
+        EasyQC2(paste0(qc2_p, "/CHCP_ADO_maxRCF_nonEUR_QC_quant.ecf"))
